[refactor] account book list와 card 분리
- 상태값으로 content, modify토글
