--- FE/src/components/AccountBook/AccountBookList/index.scss
@@ -1,119 +1,0 @@
-@use '../../../styles/values' as *;
-@use '../../../styles/mediaQuery.scss' as *;
-
-.acbook__list {
-  display: flex;
-  height: 100%;
-  width: 100%;
-  align-items: flex-end;
-  flex-direction: column;
-  gap: 3em;
-  padding: 8em 0;
-  .acbook {
-    position: relative;
-    background-color: #f9e000;
-    width: 333px;
-    min-height: 195px;
-    z-index: 1;
-
-    border-radius: 16px;
-    padding: 2em;
-    transition-duration: 0.3s;
-    transform: translateX(10%);
-    animation: cardShow 1s backwards;
-    color: black;
-
-    cursor: pointer;
-
-    .ac__title {
-      font-size: 1.5em;
-      font-weight: bold;
-      &.hidden {
-        display: none;
-      }
-    }
-    .ac__desc {
-      margin-top: 2em;
-      font-weight: 900;
-      &.hidden {
-        display: none;
-      }
-    }
-
-    .acbook__title__input {
-      position: absolute;
-      top: 2.2em;
-      height: 2em;
-      width: 230px;
-      text-decoration: none;
-
-      &.hidden {
-        display: none;
-      }
-    }
-    .acbook__desc__input {
-      position: absolute;
-      top: 5.5em;
-      height: 5em;
-      width: 230px;
-
-      resize: none;
-      &.hidden {
-        display: none;
-      }
-    }
-
-    .fas.fa-edit {
-      z-index: 2;
-      font-size: 1.3em;
-      top: 1.5em;
-      right: 1em;
-      position: absolute;
-
-      &:hover {
-        color: $spending-color;
-      }
-    }
-
-    .fas.fa-trash-alt {
-      z-index: 2;
-      font-size: 1.5em;
-      top: 3em;
-      right: 1em;
-      position: absolute;
-
-      &:hover {
-        color: $spending-color;
-      }
-    }
-    &:hover {
-      transform: translateX(-5%);
-      transition-duration: 0.3s;
-      transition-timing-function: ease, ease;
-    }
-    &:last-child {
-      margin-bottom: 3em;
-    }
-
-    @keyframes cardShow {
-      from {
-        opacity: 0;
-        transform: translateX(300px);
-      }
-
-      to {
-        opacity: 1;
-      }
-    }
-  }
-  @include mobile {
-    .acbook {
-      width: 200px;
-    }
-  }
-}
-@include mobile {
-  .acbook__list {
-    padding: 0;
-  }
-}
