+++ FE/src/components/Common/SaveModal/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import ActionButton from '../ActionButton';
-import { ResponseMessage } from '../../../util/message';
+
 import './saveModal.scss';
 
 export default function SaveModal({
