+++ FE/src/components/Common/SaveModal/saveModal.scss
@@ -2,7 +2,7 @@
 @use '../../../styles/values.scss' as *;
 
 .save__modal__overlay {
-  position: absolute;
+  position: fixed;
   width: 100%;
   height: 100%;
   display: flex;
