+++ FE/src/components/SettingContent/CalendarSetting/calendarSetting.scss
@@ -21,19 +21,3 @@
   position: relative;
   background-color: transparent;
 }
-.save__modal__overlay {
-  position: absolute;
-  width: 100%;
-  height: 100%;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  // background-color: rgba(0, 0, 0, 0.9);
-  .save__modal__content {
-    width: 50%;
-    height: 20%;
-    background-color: $modal-background;
-    border: 1px solid $dark-border-color;
-    color: $dark-general-font-color;
-  }
-}
