+++ FE/src/pages/AccountBook/index.scss
@@ -2,7 +2,7 @@
 @use '../../styles/mediaQuery' as *;
 
 .acbook__list__container {
-  position: absolute;
+  position: relative;
   width: 100%;
   height: 100%;
   display: flex;
